+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-# [0.3.0] - 2020-08-24
+# [0.3.1] - 2020-08-25
+
+### Added
+
+- `EventOf` type.
+
+### Changed
+
+- Type signature of `addEventListener` and `removeEventListener` functions.
+- Rename `AllEventMap` type to `EventMapOf`
+
+## [0.3.0] - 2020-08-24
 
 ### Added